x1plusd/expansion: refactor common LED strip logic out
